Allow the user to map joints and coordinates
Allow the user to map joints and coordinates to control busses. This allows \n the user to specify which joint of the BODY_25 \n to use as the control and which coordinate (X or Y). Due to the fact SuperCollider is awesome and already handles \n creating multiple UGen instances when calling the kr method, it's now possible to map each joint to its own control bus and have everything read out correctly \n at the same time.